Add FilteringHelper to share setup methods
